requirements.txt
# Core LangChain
langchain==1.2.10
langchain-openai>=0.1.8
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2

# OpenAI SDK (Required for LangChain 1.x)
openai>=1.0.0

# Streamlit
streamlit>=1.30.0

# Embeddings tokenizer
tiktoken>=0.5.0

# Vector store
faiss-cpu>=1.7.4

# Document loading
unstructured>=0.12.0

# Environment variables
python-dotenv>=1.0.0